Completed data synchronization

+++ composeApp/src/commonMain/kotlin/data/Sector.kt
@@ -13,7 +13,7 @@ data class Sector(
     val image: String,
     @SerialName("kids_apt") val kidsApt: Boolean,
     val weight: String,
-    @SerialName("walking_time") val walkingTime: Int? = null,
+    @SerialName("walking_time") val walkingTime: Long? = null,
     val point: LatLng? = null,
     @SerialName("sun_time") val sunTime: String,
     @SerialName("zone_id") val parentZoneId: Long,
